requirements.txt
asgiref==3.5.2
Django==4.1
django-admin-interface==0.19.2
django-ckeditor==6.5.0
django-colorfield==0.7.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
Pillow==9.2.0
six==1.16.0
sqlparse==0.4.2
